mythutil: rename commandlineparser.* to mythutil_…
