+++ mythtv/programs/mythutil/mythutil.pro
@@ -16,11 +16,11 @@ INCLUDEPATH += ../../libs/libmythbase
 QMAKE_CLEAN += $(TARGET)
 
 # Input
-HEADERS += mythutil.h commandlineparser.h
+HEADERS += mythutil.h mythutil_commandlineparser.h
 HEADERS += backendutils.h fileutils.h jobutils.h markuputils.h
 HEADERS += messageutils.h mpegutils.h musicmetautils.h
 HEADERS += recordingutils.h
-SOURCES += mythutil.cpp commandlineparser.cpp
+SOURCES += mythutil.cpp mythutil_commandlineparser.cpp
 SOURCES += backendutils.cpp fileutils.cpp jobutils.cpp markuputils.cpp
 SOURCES += messageutils.cpp mpegutils.cpp musicmetautils.cpp eitutils.cpp
 SOURCES += recordingutils.cpp
